requirements.txt
torch-cluster~=1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse~=0.6.15
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
networkx==2.5.1
pandas==1.5.1
pyyaml~=6.0
matplotlib==3.6.2
metrics==0.3.3
scikit-learn~=1.1.3
#piyaml~=0.2.5
torchmetrics~=0.10.3
#numpy~=1.23.4pip
joblib~=1.1.1